Добавлены классы

+++ task_4/BagTea.h
@@ -9,7 +9,7 @@ private:
     int gramInBag;
 
 public:
-    BagTea(int gramInBag, std::string name, std::string taste, std::string smell, int volume) : Drink(name, taste, smell, volume) {
+    BagTea(int gramInBag, std::string name, std::string taste, std::string smell, double volume) : Drink(name, taste, smell, volume) {
         this->gramInBag = gramInBag;
     }
